# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tucano Test API",
    "version": "0.1.0",
...
  },
  "servers": [
    {
      "url": "{scheme}://{host}:{port}",
      "description": "A deployment of the API. The defaults target a local `cargo run`; each variable can be overridden to point a generated client at another environment.",
      "variables": {
        "scheme": {
          "default": "http",
          "enum": [
            "http",
            "https"
          ],
          "description": "Transport scheme; use `https` when a TLS terminator fronts the service."
        },
        "host": {
          "default": "localhost",
          "description": "Hostname or address the API is reachable at."
        },
        "port": {
          "default": "3000",
          "description": "Port the API listens on."
        }
      }
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/projects": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "List projects",
        "operationId": "listProjects",
        "parameters": [
          {
            "$ref": "#/components/parameters/filter"
          },
          {
            "$ref": "#/components/parameters/tags"
          }
        ],
        "responses": {
          "200": {
            "description": "Project IDs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "$ref": "#/components/responses/Overloaded"
          },
          "504": {
            "$ref": "#/components/responses/RequestTimeout"
          }
        }
      },
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "Create a project",
        "description": "Create a project. Only a system administrator may create one, because a project that does not exist yet cannot name a grant to hold: every other authenticated caller is answered `forbidden` when authentication is enforced, and the new project starts with no grants.",
        "operationId": "createProject",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateResponse"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/BodyTooLarge"
          },
          "503": {
            "$ref": "#/components/responses/Overloaded"
          },
          "504": {
            "$ref": "#/components/responses/RequestTimeout"
          }
        }
      }
    },
    "/projects/{id}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Read a project, assembling the suites and directly owned cases it holds",
        "operationId": "getProject",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          },
          {
            "name": "children",
            "in": "query",
            "description": "Optional: `ids` answers this read’s `testSuites` and `testCases` arrays with the children’s wire identifiers instead of embedded documents (#415) — the shape a context tree fetches, priced at one folder walk. An unknown value, or the parameter absent, keeps the embedded documents; the parameter is ignored by every resource that does not hydrate children.",
            "schema": {
              "type": "string",
              "enum": [
                "ids"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project document",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidId"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Overloaded"
          },
          "504": {
            "$ref": "#/components/responses/RequestTimeout"
          }
        }
      },
      "put": {
        "tags": [
          "Projects"
        ],
        "summary": "Update a project",
        "operationId": "updateProject",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          },
          {
            "$ref": "#/components/parameters/IfMatch"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated. `document` carries the project as stored.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidIdOrRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "413": {
            "$ref": "#/components/responses/BodyTooLarge"
          },
          "503": {
            "$ref": "#/components/responses/Overloaded"
          },
          "504": {
            "$ref": "#/components/responses/RequestTimeout"
          }
        }
      },
      "delete": {
        "tags": [
          "Projects"
        ],
        "summary": "Delete a project and everything below it",
        "operationId": "deleteProject",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidId"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Overloaded"
          },
          "504": {
            "$ref": "#/components/responses/RequestTimeout"
          }
        }
      }
    },
...
    "/workflows": {
      "get": {
        "tags": [
          "Workflows"
        ],
        "summary": "List workflows (retired)",
        "description": "Workflows are created inside a project: POST /projects/{id}/workflows. This route is registered only to explain that.",
        "operationId": "listWorkflows",
        "responses": {
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "Workflows"
        ],
        "summary": "Create a workflow (retired)",
        "description": "Workflows are created inside a project: POST /projects/{id}/workflows. This route is registered only to explain that.",
        "operationId": "createWorkflow",
        "responses": {
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/workflows/{id}": {
      "get": {
        "tags": [
          "Workflows"
        ],
        "summary": "Read a workflow",
        "description": "The workflow document, resolved globally. A workflow that several projects hold is a conflict.",
        "operationId": "getWorkflow",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "The workflow document",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workflow"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Overloaded"
          },
          "504": {
            "$ref": "#/components/responses/RequestTimeout"
          }
        }
      },
      "put": {
        "tags": [
          "Workflows"
        ],
        "summary": "Update a workflow",
        "description": "Partial merge. The caller needs `editor` in the home project.",
        "operationId": "updateWorkflow",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          },
          {
            "$ref": "#/components/parameters/IfMatch"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated. `document` carries the workflow as stored.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "413": {
            "$ref": "#/components/responses/BodyTooLarge"
          },
          "503": {
            "$ref": "#/components/responses/Overloaded"
          },
          "504": {
            "$ref": "#/components/responses/RequestTimeout"
          }
        }
      },
      "delete": {
        "tags": [
          "Workflows"
        ],
        "summary": "Delete a workflow",
        "description": "The caller needs `editor` in the home project.",
        "operationId": "deleteWorkflow",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Overloaded"
          },
          "504": {
            "$ref": "#/components/responses/RequestTimeout"
          }
        }
      }
    },
    "/projects/{id}/workflows": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "List the workflows a project owns",
        "operationId": "listProjectWorkflows",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Workflow IDs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidId"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Overloaded"
          },
          "504": {
            "$ref": "#/components/responses/RequestTimeout"
          }
        }
      },
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "Create a workflow in a project",
        "description": "Creates a workflow whose home is this project. The caller needs `editor` here. Every step target must be reachable from this project; a dangling step is `invalid_request` naming the offending index.",
        "operationId": "createProjectWorkflow",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateResponse"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/BodyTooLarge"
          },
          "503": {
            "$ref": "#/components/responses/Overloaded"
          },
          "504": {
            "$ref": "#/components/responses/RequestTimeout"
          }
        }
      }
    },
    "/projects/{id}/workflows/{workflow_id}": {
      "delete": {
        "tags": [
          "Projects"
        ],
        "summary": "Delete a workflow a project owns",
        "operationId": "deleteProjectWorkflow",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          },
          {
            "$ref": "#/components/parameters/case_id"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidId"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Overloaded"
          },
          "504": {
            "$ref": "#/components/responses/RequestTimeout"
          }
        }
      }
    },
    "/projects/{id}/workflows/{workflow_id}/run": {
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "Materialise a workflow into a run",
        "description": "Creates a run from the workflow definition. Suite steps expand to their current cases; case steps are pinned as-is. The run carries sourceWorkflowId for provenance.",
        "operationId": "runFromProjectWorkflow",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          },
          {
            "$ref": "#/components/parameters/case_id"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Run name; defaults to the workflow id."
                  },
                  "projects": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    },
                    "description": "Projects the run covers; defaults to the workflow home."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Run created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateResponse"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/BodyTooLarge"
          },
          "503": {
            "$ref": "#/components/responses/Overloaded"
          },
          "504": {
            "$ref": "#/components/responses/RequestTimeout"
          }
        }
      }
    },
    "/workflows/{id}/duplicate": {
      "post": {
        "tags": [
          "Workflows"
        ],
        "summary": "Duplicate a workflow",
        "description": "Creates a copy of the workflow. The copy is refused if any step target is not reachable in the destination project.",
        "operationId": "duplicateWorkflow",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "newId": {
                    "type": "string",
                    "description": "Identifier for the copy; derived from the source when omitted."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workflow duplicated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateResponse"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/BodyTooLarge"
          },
          "503": {
            "$ref": "#/components/responses/Overloaded"
          },
          "504": {
            "$ref": "#/components/responses/RequestTimeout"
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
...
      "id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
      "filter": {
        "name": "filter",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Case-insensitive substring filter for resource IDs"
      },
      "tags": {
        "name": "tags",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Comma-separated tags; a resource matches when it carries at least one of them (case-insensitive). Resources without a tags array never match."
      },
...
    },
...
    "responses": {
...
      "RequestTimeout": {
        "description": "`request_timeout`: the request outlived `TUCANO_REQUEST_TIMEOUT_MS` (default 300000 ms; an explicit `0` disables the deadline) and the server stopped waiting for it. A write already handed to the storage layer finishes atomically regardless of the cutoff, so a `504` answers 'unknown', never 'half done'.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/XRequestId"
          }
        }
      },
      "Overloaded": {
        "description": "`service_unavailable`: the request arrived while `TUCANO_MAX_CONCURRENCY` (default 128; an explicit `0` removes the cap) others were already in flight. The server refuses rather than queues, and the answer carries `Retry-After`.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/XRequestId"
          },
          "Retry-After": {
            "$ref": "#/components/headers/RetryAfter"
          }
        }
      },
      "Unauthorized": {
        "description": "`missing_token`, `invalid_token` or `token_expired`: the `Authorization` header was absent or carried a token this deployment does not accept. The response carries the `WWW-Authenticate` challenge.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/XRequestId"
          },
          "WWW-Authenticate": {
            "$ref": "#/components/headers/WWWAuthenticate"
          }
        }
      },
...
    },
    "headers": {
      "XRequestId": {
        "description": "The id the API assigned to this request, or the inbound `X-Request-Id` it honoured. The same value is carried as `requestId` in the JSON error envelope, so quote it when reporting a failure.",
        "schema": {
          "type": "string"
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "A signed access token carried as `Authorization: Bearer <token>`. Obtain one from `/auth/login` (or by exchanging a refresh token at `/auth/refresh`) and send it on every operation that names a caller. The token is stateless: its signature and expiry are the whole check, so authority changed after it was minted is not re-read until the token expires."
      }
    }
  }
}